requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
actuator-msgs==0.0.1
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
anyio==4.9.0
asyncio-mqtt==0.16.2
attrs==25.3.0
bond==3.0.2
builtin-interfaces==1.2.1
cartographer-ros-msgs==2.0.9002
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
composition-interfaces==1.2.1
control-msgs==4.7.0
controller-manager==2.46.0
controller-manager-msgs==2.46.0
cryptography==45.0.5
cv-bridge==3.2.1
demo-nodes-py==0.20.5
Deprecated==1.2.18
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
distro==1.9.0
domain-coordinator==0.10.0
dwb-msgs==1.1.17
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.3.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
gps-msgs==2.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
jiter==0.10.0
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
mcp==1.12.2
message-filters==4.3.5
my-first-package==0.0.0
nav-2d-msgs==1.1.17
nav-msgs==4.2.4
nav2-common==1.1.17
nav2-msgs==1.1.17
nav2-simple-commander==1.0.0
openai==1.98.0
osrf-pycommon==2.0.2
paho-mqtt==2.1.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-qt-binding==1.1.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
referencing==0.36.2
requests==2.32.4
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.16
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2controlcli==2.46.0
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.26.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
slam-toolbox==2.6.9
smclib==3.0.2
sniffio==1.3.1
sros2==0.10.5
sse-starlette==3.0.2
starlette==0.47.2
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.9
topic-monitor==0.20.5
tqdm==4.67.1
trajectory-msgs==4.2.4
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
typing-inspection==0.4.1
typing_extensions==4.14.1
unique-identifier-msgs==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
vision-msgs==4.1.1
visualization-msgs==4.2.4
wrapt==1.17.2
xacro==2.0.8
google-generativeai
